Finish up file move and building
