+++ src/Servers/IIS/IIS/test/Common.FunctionalTests/LoggingTests.cs
@@ -13,7 +13,7 @@ using Xunit;
 namespace Microsoft.AspNetCore.Server.IIS.FunctionalTests
 {
     [Collection(PublishedSitesCollection.Name)]
-    public class LoggingTests : LogFileTestBase
+    public class LoggingTests : IISFunctionalTestBase
     {
         public LoggingTests(PublishedSitesFixture fixture) : base(fixture)
         {
@@ -50,7 +50,7 @@ namespace Microsoft.AspNetCore.Server.IIS.FunctionalTests
         private async Task CheckStdoutToFile(TestVariant variant, string path)
         {
             var deploymentParameters = Fixture.GetBaseDeploymentParameters(variant);
-            deploymentParameters.EnableLogging(_logFolderPath);
+            deploymentParameters.EnableLogging(LogFolderPath);
 
             var deploymentResult = await DeployAsync(deploymentParameters);
 
@@ -58,7 +58,7 @@ namespace Microsoft.AspNetCore.Server.IIS.FunctionalTests
 
             StopServer();
 
-            var contents = Helpers.ReadAllTextFromFile(Helpers.GetExpectedLogName(deploymentResult, _logFolderPath), Logger);
+            var contents = Helpers.ReadAllTextFromFile(Helpers.GetExpectedLogName(deploymentResult, LogFolderPath), Logger);
 
             Assert.Contains("TEST MESSAGE", contents);
             Assert.DoesNotContain("\r\n\r\n", contents);
@@ -192,7 +192,7 @@ namespace Microsoft.AspNetCore.Server.IIS.FunctionalTests
             var deploymentParameters = Fixture.GetBaseDeploymentParameters(Fixture.InProcessTestSite, variant.HostingModel);
             deploymentParameters.TransformArguments((a, _) => $"{a} {path}"); // For standalone this will need to remove space
 
-            var logFolderPath = _logFolderPath + "\\彡⾔";
+            var logFolderPath = LogFolderPath + "\\彡⾔";
             deploymentParameters.EnableLogging(logFolderPath);
 
             var deploymentResult = await DeployAsync(deploymentParameters);
@@ -215,14 +215,14 @@ namespace Microsoft.AspNetCore.Server.IIS.FunctionalTests
         {
             var deploymentParameters = Fixture.GetBaseDeploymentParameters(variant);
 
-            deploymentParameters.EnableLogging(_logFolderPath);
+            deploymentParameters.EnableLogging(LogFolderPath);
 
             var deploymentResult = await DeployAsync(deploymentParameters);
             await Helpers.AssertStarts(deploymentResult, "ConsoleWrite");
 
             StopServer();
 
-            Assert.Single(Directory.GetFiles(_logFolderPath), Helpers.GetExpectedLogName(deploymentResult, _logFolderPath));
+            Assert.Single(Directory.GetFiles(LogFolderPath), Helpers.GetExpectedLogName(deploymentResult, LogFolderPath));
         }
 
         [ConditionalFact]
@@ -270,6 +270,31 @@ namespace Microsoft.AspNetCore.Server.IIS.FunctionalTests
             StopServer();
 
             EventLogHelpers.VerifyEventLogEvent(deploymentResult, EventLogHelpers.OutOfProcessFailedToStart(deploymentResult, new string('a', 30000)), Logger);
+        }
+
+        [ConditionalTheory]
+        [MaximumOSVersion(OperatingSystems.Windows, WindowsVersions.Win10_20H1, SkipReason = "Shutdown hangs https://github.com/dotnet/aspnetcore/issues/25107")]
+        [InlineData("ConsoleErrorWriteStartServer")]
+        [InlineData("ConsoleWriteStartServer")]
+        public async Task CheckStdoutLoggingToPipeWithFirstWrite(string path)
+        {
+            var deploymentParameters = Fixture.GetBaseDeploymentParameters();
+
+            var firstWriteString = "TEST MESSAGE";
+
+            deploymentParameters.TransformArguments((a, _) => $"{a} {path}");
+
+            var deploymentResult = await DeployAsync(deploymentParameters);
+
+            await Helpers.AssertStarts(deploymentResult);
+
+            StopServer();
+
+            if (deploymentParameters.ServerType == ServerType.IISExpress)
+            {
+                // We can't read stdout logs from IIS as they aren't redirected.
+                Assert.Contains(TestSink.Writes, context => context.Message.Contains(firstWriteString));
+            }
         }
 
         private static string ReadLogs(string logPath)
