--- src/Servers/IIS/IIS/test/Common.FunctionalTests/ServerAbortTests.cs
@@ -1,21 +1,0 @@
-// Copyright (c) .NET Foundation. All rights reserved.
-// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
-
-using System.Net;
-using System.Net.Http;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Server.IIS.FunctionalTests.Utilities;
-using Microsoft.AspNetCore.Server.IntegrationTesting;
-using Microsoft.AspNetCore.Testing;
-using Xunit;
-
-namespace Microsoft.AspNetCore.Server.IIS.FunctionalTests
-{
-    [Collection(PublishedSitesCollection.Name)]
-    public class ServerAbortOutOfProcessTests : IISFunctionalTestBase
-    {
-        public ServerAbortOutOfProcessTests(PublishedSitesFixture fixture) : base(fixture)
-        {
-        }
-    }
-}
